requirements.txt
biothings_client==0.2.6
tqdm==4.64.0
requests==2.24.0
pandas==1.0.5
numpy==1.22.4
rdkit-pypi==2022.3.2.1
rdflib==6.1.1
sklearn==1.1.1
xgboost==1.3.3
scipy==1.7.0
IPython==7.16.1
pyrdf2vec==0.2.3
flask==1.1.2
flask_wtf==1.0.1
wtforms==3.0.1
nest-asyncio
gensim
aiohttp
